backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
torch==2.8.0
torchvision==0.23.0
open-clip-torch==2.24.0
Pillow==10.1.0
numpy==1.26.4
scikit-learn==1.3.2
opencv-python==4.8.1.78
requests==2.31.0
aiofiles==23.2.1
boto3==1.34.0
minio==7.2.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
openai==1.3.7
httpx==0.25.2
transformers==4.44.0
pytest==7.4.3
pytest-asyncio==0.21.1

# Enhanced ML dependencies
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
xgboost==2.0.1
lightgbm==4.1.0
optuna==3.4.0
wandb==0.16.0
tensorboard==2.15.1
plotly==5.17.0
pandas==2.2.2
scipy==1.11.4

# Additional AI/ML Enhancement Libraries
sentence-transformers==2.2.2
Flask==3.0.0
Flask-Cors==4.0.0
nltk==3.8.1
jieba==0.42.1
jsonlines==4.0.0
marshmallow==3.20.1
ngrok==1.0.0
peft==0.6.2
psutil==5.9.6
rich==13.7.0
simhash==2.1.2
tiktoken==0.5.2
jinja2==3.1.2
trl==0.7.4
ujson==5.8.0
streamlit==1.28.2
datasets==2.14.6
datasketch==1.6.4